Tuned the uint_64bit type.

+++ hpcopr/now_md5.h
@@ -11,7 +11,7 @@
 typedef unsigned char uint_8bit;
 typedef unsigned short uint_16bit;
 typedef unsigned int uint_32bit;
-typedef unsigned long int uint_64bit;
+typedef unsigned long long int uint_64bit;
 
 #define rot_left(a,n) (((a)<<(n))|((a)>>(32-(n))))
 #define F(b,c,d) (((b)&(c))|((~b)&(d)))
